refactor: input state type 분리
