+++ src/Input/textarea/textarea.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, Suspense, useEffect, useState } from "react";
 import ErrorIcon from "../../icon/svg/status/warning-circle-filled.svg";
 import WarnIcon from "../../icon/svg/status/warning-triangle-filled.svg";
 import TextAreaSkeleton from "./textareaSkeleton";
-import { InputStateType } from "../textfield/textfield";
+import { InputStateType } from "src/utils/type";
 
 type TextFieldProps = {
   value: string;
