+++ src/utils/type.ts
@@ -1,0 +1,6 @@
+export type InputStateType =
+  | "enable"
+  | "warning"
+  | "error"
+  | "disabled"
+  | "readOnly";
